增加玩家基本信息改变推送 增加dto返回客户端所需数据 使用java8重构背包 去掉缓存，方便调试

+++ src/main/java/com/linlazy/mmorpglintingyang/utils/ItemIdUtil.java
@@ -28,8 +28,8 @@ public class ItemIdUtil {
      * @param baseItemId
      * @return
      */
-    public static long getNewItemId(int orderId, int backPackIndex, int baseItemId) {
-        int i = (orderId << 40) + (backPackIndex << 28) + baseItemId;
+    public static long getNewItemId(long orderId, int backPackIndex, int baseItemId) {
+        long i = (orderId << 40) + (backPackIndex << 28) + baseItemId;
         return i;
     }
